+++ doc/CHANGELOG.md
@@ -32,7 +32,9 @@ For release instructions, see [cynic-net/pypi-release] on GitHub.
 - Changed: new toolset.bm2 download location
 - Added: `t8dev emulate runcpm` now copies files to the "disks" and can
   optionally auto-run the first file given.
-- Added: Optional `addr` param for `testmc.generic.machine.stepto()`.
+- Added: In `testmc.generic.machine`, `stepto()` now has an optional `addr`
+  (start address) parameter, and `stepto()` and `call()` have an optional
+  `nstop` (number of times to see a stop point before returning) parameter.
 
 ### 0.2.0 (2024-11-04)
 - API: `testmc.*.Machine.call()` now executes return instructions, rather